server/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Flask==1.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
mariadb==1.0.8
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.28
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
